requirements.txt
python-dotenv
uptime-kuma-api
pandas